Removed ContainerDeserializer interface.
